restaurant-lookup-service/requirements.txt
boto3>=1.28.0
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=1.4.0
asyncpg>=0.27.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=1.10.7